[DD-1291] Rework copyright script to update existing headers
This also fixes some issues along the way. The check if part of a file is a copyright banner was broken in many ways (each bug was unfolded after the previous one was fixed; they have been hiding each other, nasty little bugs):
- using the wrong lodash function (maybe due to update) – no lines were even checked in the first place, marking every file as 'ok' regardless of their content (it was actually this bug, not the [one I suspected in the mail](https://github.com/dswarm/dswarm-backoffice-web/blob/5ea32b7e35de058060b864486cad4daf1ef0fcb8/yo/copyright.js#L97-L99) that was responsible for the issue)
- only checking against `'copyright'` was to broad, now the line must end with whatever copyright line would be generated based on the config (it's the line that changes based on the year)
- second, lines were checked sensitive to whitespace at the end which some editors remove. This meant that the header would have been partially deleted and then written completely new, rendering the file unusable (syntactically wrong)
- third, everything was included in the banner as long as it matches any of the special comment markers. For `less` files, this is an empty string, so the checker would mark __everything__ in a less file as old copyright and subsequently remove it.
This is now reworked to use a state-machine like checker that requires 1 start line, a number of middle lines that might be right trimmed and one end line. This works for all our cases.
In addition to the actual feature, that existing copyright banners could be updated, I made everything work in a streaming sort-of way and removed the update task, that we didn't use and is superseded by ensure.

Signed-off-by: Paul Horn <phorn@avantgarde-labs.de>